# saas-app-ui/src/app/api-services/models/page-response-category-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CategoryResponse } from '../models/category-response';
export interface PageResponseCategoryResponse {
  content?: Array<CategoryResponse>;
  first?: boolean;
  hasNext?: boolean;
  hasPrevious?: boolean;
  last?: boolean;
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}