# saas-app/src/main/java/com/amoutech/saas/responses/CategoryResponse.java
package com.amoutech.saas.responses;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class CategoryResponse {

    private String id;
    private String name;
    private String description;
    private int nbProducts;
}