# saas-app/src/main/java/com/amoutech/saas/responses/StockMvtResponse.java
package com.amoutech.saas.responses;

import com.amoutech.saas.entities.TypeMvt;
import lombok.*;

import java.time.LocalDate;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class StockMvtResponse {

    private String id;
    private TypeMvt typeMvt;
    private Integer quantity;
    private LocalDate dateMvt;
    private String comment;
}